requirements.txt
fastapi
uvicorn
google-generativeai
jinja2
python-multipart
python-dotenv
gunicorn